feat(ai-cli): add support for Google Gemini Model
This commit adds support for the Google Gemini Model to the `ai-cli` tool.

The Gemini Model is a large language model developed by Google AI. It is similar to the OpenAI GPT-3 model, but it is trained on a larger dataset and has access to real-time information.

To use the Gemini Model, you can set the `GOOGLE_API_KEY` environment variable or specify it using the `--google-api-key` parameter. You can also set it using the `ai setting` command. Need to set bot to `GeminiBot` by `ai setting --edit bot=GeminiBot`.